backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
python-decouple==3.8
Pillow==10.1.0
opencv-python==4.9.0.80
mediapipe==0.10.9
celery==5.3.4
redis==5.0.1
requests==2.31.0
python-multipart==0.0.6
gunicorn==21.2.0
google-generativeai>=0.3.0
psycopg2-binary>=2.9.7
dj-database-url>=2.1.0
whitenoise>=6.6.0 